+++ test/application.test.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-unused-expressions */
+/* eslint-disable no-undef */
 import '../node_modules/regenerator-runtime/runtime.js';
 import '../dist/snjs.js';
 import '../node_modules/chai/chai.js';
@@ -15,7 +17,7 @@ describe('application instances', () => {
 
   after(async () => {
     localStorage.clear();
-  })
+  });
 
   it('two distinct applications should not share model manager state', async () => {
     const app1 = await Factory.createAndInitializeApplication('app1');
@@ -23,7 +25,7 @@ describe('application instances', () => {
     expect(app1.modelManager).to.equal(app1.modelManager);
     expect(app1.modelManager).to.not.equal(app2.modelManager);
 
-    const app1Item = await Factory.createMappedNote(app1);
+    await Factory.createMappedNote(app1);
     expect(app1.modelManager.allItems.length).length.to.equal(BASE_ITEM_COUNT + 1);
     expect(app2.modelManager.allItems.length).to.equal(BASE_ITEM_COUNT);
   });
@@ -32,13 +34,13 @@ describe('application instances', () => {
     const app1 = await Factory.createAndInitializeApplication('app1');
     const app2 = await Factory.createAndInitializeApplication('app2');
 
-    const app1Item = await Factory.createMappedNote(app1);
+    await Factory.createMappedNote(app1);
     await app1.syncManager.sync();
 
     expect((await app1.storageManager.getAllRawPayloads()).length).length.to.equal(BASE_ITEM_COUNT + 1);
     expect((await app2.storageManager.getAllRawPayloads()).length).length.to.equal(BASE_ITEM_COUNT);
 
-    const app2Item = await Factory.createMappedNote(app2);
+    await Factory.createMappedNote(app2);
     await app2.syncManager.sync();
 
     expect((await app1.storageManager.getAllRawPayloads()).length).length.to.equal(BASE_ITEM_COUNT + 1);