+++ test/sync_tests/notes_tags.test.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-unused-expressions */
+/* eslint-disable no-undef */
 import '../../node_modules/regenerator-runtime/runtime.js';
 import '../../dist/snjs.js';
 import '../../node_modules/chai/chai.js';
@@ -16,8 +18,8 @@ describe("notes + tags syncing", async function() {
 
   beforeEach(async function() {
     this.application = await Factory.createInitAppWithRandNamespace();
-    const email = SFItem.GenerateUuidSynchronously();
-    const password = SFItem.GenerateUuidSynchronously();
+    const email = Uuid.GenerateUuidSynchronously();
+    const password = Uuid.GenerateUuidSynchronously();
     await Factory.registerUserToApplication({application: this.application, email, password});
   })